Warnings as errors defaults to true. Gpu profiling PLATFORM_DEBUG_PROFILING_LEVELS CMake option added. This can help to figure out mismatched Begin()/End() calls, but will be slow if activated.

+++ DirectX12/VideoDecoder.cpp
@@ -173,7 +173,7 @@ cp::VideoDecoderResult VideoDecoder::DecodeFrame(cp::Texture* outputTexture, con
 		return cp::VideoDecoderResult::InvalidDecodeArgumentType;
 	}
 
-	inputArgs.ReferenceFrames.NumTexture2Ds = mTextures.size();
+	inputArgs.ReferenceFrames.NumTexture2Ds = (uint)mTextures.size();
 
 	for(int i = 0; i < mTextures.size(); ++i)
 	{
